Simplify enrollment default index; drop extra module and tests
Keep matching in enrollmentHistory.ts only: newest-first scan for same
term, then same section code, then instructor overlap, else newest row.
Remove enrollmentHistoryPopoverMatch and the Vitest file per request.
